Move temp_dir with context to silx.testutils
